Feature Update
1. Rearrange the data structure of GPA.json and related .qml
2. Add functions to write record into GPA.json
3. Modify some function prototype in MyListModel to integrate with
RecordModel
4. Add condition judgment before writing new record

Known bugs:
1. need to adjust the order of r/w data from/to GPA.json file
2. need to be more specific on the record time

+++ src/gpaModel.cpp
@@ -1,6 +1,6 @@
 #include "gpaModel.h"
 
-float computeCGPA(float currentGpa, float unitsNumberPassed, QVariantList  grades , QVariantList courseUnits , int size){
+double computeCGPA(float currentGpa, float unitsNumberPassed, QVariantList  grades , QVariantList courseUnits , int size){
 	float totalUnits(0) , totalGradePoints(0) , termGPA(0) , cGPA(0);
 
 
@@ -15,7 +15,7 @@ float computeCGPA(float currentGpa, float unitsNumberPassed, QVariantList  grade
 
 }
 
-float computeGradePoints(QVariant c){
+double computeGradePoints(QVariant c){
 	float gradePoint(0) ;
 
 	if (c=="A+" || c=="a+"){
